requirements.txt
absl-py==0.7.1
astor==0.8.0
attrs==19.1.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
DateTime==4.3
Fiona==1.8.6
gast==0.2.2
geopandas==0.5.0
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lightgbm==2.2.3
Markdown==3.1.1
matplotlib==3.1.0
mock==3.0.5
munch==2.3.2
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
pyparsing==2.4.0
pyproj==2.1.3
python-dateutil==2.8.0
python-geohash==0.8.5
pytz==2019.1
PyYAML==5.1.1
scikit-learn==0.21.2
scipy==1.3.0
Shapely==1.6.4.post2
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.4
xgboost==0.90
zope.interface==4.6.0
